feat: Implement Employee Check-In/Check-Out functionality with location confirmation and history tracking
