+++ app/(tabs)/home.tsx
@@ -1,4 +1,4 @@
-import EmpCheckIn from "@/components/EmpCheckIn";
+import EmpCheckIn from "@/components/EmployeeCheckIn/EmpCheckIn";
 import { useTheme } from "@/contexts/ThemeContext";
 import { Ionicons } from "@expo/vector-icons";
 import {
